WIP: move client calls to `useDesktopSession`
this isn't fully complete but it gets the idea across.
`useTdpClientCanvas` shouldn't set anything when it comes to client
callbacks. Really, the only thing `useTdpClientCanvas` should be is a
glorified object of canvas callbacks that we can spread as props into
TdpClientCanvas.

From the looks of it, we still need it in `useDesktopSession` because of
the canvasOnPngFrame and canvasOnnBmpFrame methods that call a draw to
the canvas. We basically return those methods from `useTdpClientCanvas`
and then call them in the `useDesktopSession` `onPngFrame` etc.

That should be just about it.

Will return to it on Monday but I think its getting close (as a
refactor, then we can make any other cleanups later)